vajbcha/requirements.txt
black==26.3.1
blinker==1.9.0
click==8.3.2
Flask==3.1.3
isort==8.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
librt==0.9.0
MarkupSafe==3.0.3
mypy==1.20.0
mypy_extensions==1.1.0
packaging==26.0
pathspec==1.0.4
pillow==12.2.0
platformdirs==4.9.6
pur==7.3.3
pytokens==0.4.1
typing_extensions==4.15.0
Werkzeug==3.1.8
